Changes to Event Info
Updated buttons to be orange. Made remove event pop-up smaller, Fixed text sizing on some buttons.

Made it so that fields that cannot be editable are greyed out. I.e in the event of an actual reservation, some info cannot be changed. The app will need to track if an event was booked or created,